flask-app/requirements.txt
# DB
sqlalchemy
geoalchemy2

# Flask
flask
flask_marshmallow
flask_sqlalchemy
marshmallow_sqlalchemy
flask-cors

# Geometry
shapely